feat: move idn_connect, fix feature gating
