+++ lib/Cargo.toml
@@ -45,7 +45,6 @@ ark-serialize = "0.4.0"
 w3f-bls = "0.1.3"
 murmur-core = { path = "../core/", features = ["client"] }
 zeroize = "1.8.1"
-hkdf = "0.12.4"
 
 
 [dev-dependencies]
